linear_regression/multiple_non_lib/requirements.txt
numpy==1.11.3
pandas==0.22.0
python-sklearn